### 2026-09-04 #11251 @editorialbot
[PRE REVIEW]: TanML: Automated Model Validation Toolkit for Tabular Machine Learning Models
**Submitting author:** @tradertanmay (<a href="http://orcid.org/0009-0004-8583-2208">Tanmay Sah</a>)
**Repository:** https://github.com/tdlabs-ai/tanml
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.2.2
