### 2026-09-04 #11253 @crvernon
This submission has already been reviewed and accepted by pyOpenSci (https://github.com/pyOpenSci/software-submission/issues/280) and is in scope for JOSS, so its publication here will be fast tracked.
