### 2026-09-04 #11253 @crvernon
@tradertanmay - please address the following before I accept this submission for publication.

- [ ] TABLE 1: Remove the extra space before the comma in “Data, Model , Governance”
- [ ] LINE 80: Change “Core Architecture” to sentence case (“Core architecture”) for consistency with headings such as “Software design,” “Statement of need,” and “State of the field.”
- [ ] LINE 100: Change “Modular Workflow” to sentence case (“Modular workflow”) for consistency with the other section headings.
- [ ] LINE 139-140: “Software Design section” should match the actual section title capitalization, i.e., “Software design section.”
- [ ] LINE 168: Remove the erroneous word “slogans” from the author list: “Ke, G., Meng, Q. slogans, Finley...” should be corrected to the proper author list.  Also the DOI on this is a dead link.
- [ ] LINE 181-183: The “p” in “python” should be capitalized: “Machine learning in Python.”
- [ ] LINE 190-191: The “p” in “python” should be capitalized: “modeling with Python.”
- [ ] LINE 193: Standardize the software name capitalization. “Ydata-profiling” is inconsistent with “YData Profiling” used on line 47.
- [ ] LINE 194-195: The Yurdakul reference lacks enough publication information to be readily retrievable. Add the document type and/or repository, URL, DOI, or other persistent identifier if one exists.

Let me know when you are finished.  Thank you.
