### 2026-09-04 60f6b58
docs(joss): address editor feedback on citations, capitalization, and table formatting
