+++ paper.bib
@@ -9,7 +9,7 @@
 
 @article{pedregosa2011scikit,
   author  = {Pedregosa, Fabian and Varoquaux, Ga{\"e}l and Gramfort, Alexandre and Michel, Vincent and Thirion, Bertrand and Grisel, Olivier and Blondel, Mathieu and Prettenhofer, Peter and Weiss, Ron and Dubourg, Vincent and VanderPlas, Jake and Passos, Alexandre and Cournapeau, David and Brucher, Matthieu and Perrot, Matthieu and Duchesnay, {\'E}douard},
-  title   = {Scikit-learn: Machine Learning in Python},
+  title   = {Scikit-learn: Machine Learning in {P}ython},
   journal = {Journal of Machine Learning Research},
   year    = {2011},
   volume  = {12},
@@ -29,12 +29,11 @@
 }
 
 @inproceedings{ke2017lightgbm,
-  author    = {Ke, Guolin and Meng, Qi slogans and Finley, Thomas and Wang, Taifeng and Chen, Wei and Ma, Weidong and Ye, Qiwei and Liu, Tie-Yan},
+  author    = {Ke, Guolin and Meng, Qi and Finley, Thomas and Wang, Taifeng and Chen, Wei and Ma, Weidong and Ye, Qiwei and Liu, Tie-Yan},
   title     = {LightGBM: A Highly Efficient Gradient Boosting Decision Tree},
   booktitle = {Advances in Neural Information Processing Systems 30},
   year      = {2017},
-  doi       = {10.5555/3294996.3295074},
-  url       = {https://papers.nips.cc/paper/6907-lightgbm-a-highly-efficient-gradient-boosting-decision-tree}
+  url       = {https://proceedings.neurips.cc/paper_files/paper/2017/hash/6449f44a102fde848669bdd9eb6b76fa-Abstract.html}
 }
 
 @inproceedings{prokhorenkova2018catboost,
@@ -56,7 +55,7 @@
 
 @software{clemente2023ydata,
   author   = {{YData}},
-  title    = {ydata-profiling},
+  title    = {{YData Profiling}},
   year     = {2026},
   doi      = {10.5281/zenodo.10123516},
   url      = {https://github.com/ydataai/ydata-profiling},
@@ -148,22 +147,22 @@
   publisher = {IEEE Computer Society}
 }
 
-@article{seabold2010statsmodels,
-  title={Statsmodels: econometric and statistical modeling with python.},
-  author={Seabold, Skipper and Perktold, Josef and others},
-  journal={scipy},
-  volume={7},
-  number={1},
-  pages={92--96},
-  year={2010}
+@inproceedings{seabold2010statsmodels,
+  title     = {Statsmodels: Econometric and Statistical Modeling with {P}ython},
+  author    = {Seabold, Skipper and Perktold, Josef},
+  booktitle = {Proceedings of the 9th Python in Science Conference},
+  pages     = {92--96},
+  year      = {2010},
+  url       = {https://conference.scipy.org/proceedings/scipy2010/pdfs/statsmodels.pdf}
 }
 
-
-@book{yurdakul2018statistical,
-  title={Statistical properties of population stability index},
-  author={Yurdakul, Bilal},
-  year={2018},
-  publisher={Western Michigan University}
+@phdthesis{yurdakul2018statistical,
+  title  = {Statistical Properties of the Population Stability Index},
+  author = {Yurdakul, Bilal},
+  year   = {2018},
+  school = {Western Michigan University},
+  type   = {Doctoral dissertation},
+  url    = {https://scholarworks.wmich.edu/dissertations/3208}
 }
 
 
