### 2026-09-04 eb56938
docs: add JOSS publication badge, update CITATION.cff, and add JOSS citation
