### 2026-09-04 b08cc1d
docs: sync CITATION.cff and README with JOSS badge from main
