### 2026-09-04 #11253 @tradertanmay
Thank you so much to the editorial team @crvernon for the smooth review and for publishing our paper in JOSS!
