### 2026-09-04 #11253 @crvernon
🥳 Congratulations @tradertanmay on your paper being published in JOSS!  Thanks to the pyOpenSci editor @crhea93 and reviewers @echen1214 and @anabeltan for the main review.
